Fix provided mods.
The issue here was the `ModDep` would not resolve the true "root" of the load option.

A load option is a mod that can or may be loaded. The issue here was that a provided mod load option (`ProvidedModOption`) was added to the dependency helper rather than the root, which represents the mod providing the provided mod. Because of this bug, the provided mod option was not considered as a duplicate, which meant that the provided mod and real mod were considered to be part of the compatable mod set. This means when addMod is called, the candidate in the load option for the real mod was added twice.