Fix provided mods. (#24)
* Fix provided mods.

From Alex: Technically the reason is that we don't want to have multiple "variables" for a single mod - as a provided mod is always present if the providing mod is present, and always missing if the providing mod is not present. Using getRoot ensures we actually depend on the provider (and thus we solve correctly) rather than potentially saying that a "provided" mod is present but the provider isn't (or vice versa)

* Use array for allowed load options in ModDep#put 

Ran again with a mod depending on a provided mod to ensure this still works. It does.