Replace JIMFS with our own (much smaller) impl.
Currently this has an (unused) ReadOnly filesystem too, but maybe we'll try using that later on? (Possibly in the plugin branch? Either way it's only here to make merging simpler later on).

Only tested in a dev environment, hence why it's not on master.

This doesn't have *any* other changes - this is purely a JIMFS -> QuiltMemeoryFileSystem conversion.