Real, actual (and tested against our internal tests) fix for #65.
My previous commit was close, but didn't take into account that providing a mod technically
counted as a separate mod from OptionalModIdDefinition's persepective. This caused solving
to fail when a mod provided the same mod id multiple times - possibly with multiple different
groups, or the same mod id as itself. The fix is to de-duplicate those providing mods with
the real mods, which ensures that a single mod doesn't count multiple times to the "up to 1"
mod per ID limit.