Change how mismatched environment dependencies are handled, which restores the previous behaviour.
Ideally  we would add "environment" to a depends block, and then require every mod to declare it for sided mods - however we can't mandate that existing mods do that, so this workaround is necessary.

(It's still going to be useful to declare and read it though - since that will allow a sided mod to be missing when on the wrong side).