Ignore duplicate mods of the same version in development environments.
Quilt Loader 0.17 and Fabric Loader does this in all environments, but I don't like
that at all. However, many complicated, multi-mod dev environments (QSL, QFAPI) depend
on this behavior. While it is probably possible to fandangle a buildscript which does
not put the same mod on the classpath multiple times, this is a much simpler solution
that matches generally expected behavior.

Co-authored-by: AlexIIL <AlexIIL@users.noreply.github.com>