Include the mod file name in caches.
This is to help debug why a mods hash changed, as I suspect most of the time it will be due to *updating* the mod rather than the mod file changing in any other way.
Repeated cache changes could also be caused by something going wrong with how loader calculates the hash though.