Port to mapping-io (Glitch's Version) (#432)
* Port to mapping-io (Glitch's Version)

This is step 0 in adding the ability for Loader to use alternative intermediates for mods, or use something like mojmaps at runtime. This commit migrates the pre-existing MappingConfiguration to use mapping-io, but doesn't do much to actually refactor Loader to take advantage of mapping-io's strengths.

This PR moves a few dependencies to be shadowed that Floader moved as well. Tiny Remapper was required to be moved to work with mapping-io, tiny mapping parser was removed by upstream, and I just decided to bring AccessWidener in line with upstream as well.

(PS: I know QMT exists-ish, but it's not finished and I want to work on a modloader, not a mappings parser. When and if QMT is finished, we can evaluate migrating to it, but the potential to use a first-party library later shouldn't prevent us from improving Loader today)

* Fix "duplicate class" spam from Tiny Remapper in RuntimeModRemapper

TR eagerly opens nested ZIP files and flattens them, and this slightly bizarre behavior
causes lots of issues when trying to remap mods that include JiJ candidates. To mitigate this,
we create a read-only view of *only* the `.class` files. This has minimal memory impact as the
FS is copy-on-write, so it is just a big map of file paths.

* Update actions for Java 21

* Update actions for Java 21 take 2

* Mount instead of CoW

* Remove Proguard

Proguard is magically adding SequencedCollection to the LVT of MemoryMappingTree.accept
in mapping-io. We couldn't find a minimal reproduction, so we're just going to move away from
it for now.

* Fix mods that need remapping, but don't contain any class files, causing the remapper to crash because the mod directory didn't exist.

Also comment out the DedicatedServerOnly annotation on TestMixinGuiHelper, because we're not testing package-stripping annotations any more. This allows us to run the test mod.

* Mount temporary files as READ_ONLY

---------

Co-authored-by: AlexIIL <alexj9@me.com>
Co-authored-by: AlexIIL <AlexIIL@users.noreply.github.com>